21_08_2024_asyncio_aiorgam_sqlalchemy_request/requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncmy==0.2.9
attrs==24.1.0
certifi==2024.7.4
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
yarl==1.9.4


requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
mysql-connector-python~=9.0.0
matplotlib~=3.9.2
pillow~=10.4.0